fix warnings: unused variables and functions
CLang noticed several variables and functions that weren't being used anymore,
and raised warnings for them, like:

    warning: unused variable 'n_population' [-Wunused-variable]
    warning: unused function 'validDifficulty' [-Wunused-function]
    warning: private field 'caller_button' is not used [-Wunused-private-field]

Most of the the warnings were spot on, but there were a couple that weren't
*quite* so simple.

validDifficulty was defined in game.h, but only called from one unit. This
patch moves validDifficulty() from the header to the calling unit, next to
validPlayer().

The OneMouseButonPanel class declared a private field `caller_button` which
CLang thought was unused. Although caller_button is passed in to
OneMouseButtonPanel constructors and methods, the value is used immediately
(not stored in a field).